src/requirements.txt
py-cord==2.6.1
python-dotenv==1.0.1
google-genai==1.7.0
PyNaCl==1.5.0
pytz==2025.2